api_gateway/requirements.txt
django==5.1
django-cors-headers==4.4
requests==2.32
python-dotenv==1.0.1
gunicorn==22.0
